move method configuration to new service
