+++ core/src/main/scala/org/broadinstitute/dsde/rawls/util/WorkspaceSupport.scala
@@ -4,7 +4,7 @@ import akka.http.scaladsl.model.StatusCodes
 import cats.implicits.{catsSyntaxApplyOps, toFoldableOps}
 import cats.ApplicativeThrow
 import org.broadinstitute.dsde.rawls._
-import org.broadinstitute.dsde.rawls.dataaccess.{SamDAO, SlickDataSource}
+import org.broadinstitute.dsde.rawls.dataaccess.SamDAO
 import org.broadinstitute.dsde.rawls.model.{
   ErrorReport,
   RawlsRequestContext,
